PAP-2075 centralize active-run watchdog recovery
